Move most of matching_results out; move search_configs out
